Remove unnecessary test
The `patfinder` executable behaves differently now, it always returns a
transferSteps path as soon as there is any max flow given.

The expected behaviour now is that the `transfer` method will throw an
exception as soon as the max flow is lower than the transfer value. This
is covered by another test already.